feat: talking system done
