--- Erlang-Legacy/Assets/Scripts/Dialogue.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f9e671f56adb6504faf00ff97f1185fa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
