--- Erlang-Legacy/Assets/Scripts/DialogueManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ed0495ddb4e708d44ab0efa8443a3252
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
