+++ Erlang-Legacy/Assets/Scripts/GameSession/SceneEntrance.cs
@@ -1,7 +1,6 @@
 ﻿using Core.Player.Controller;
-using Core.Player.Util;
 using Core.Shared.Enum;
-using DG.Tweening;
+using Core.Shared;
 using UnityEngine;
 
 namespace Core.GameSession
@@ -24,44 +23,20 @@ namespace Core.GameSession
         //post: animates player entering the current scene
         public void MakeEntrance()
         {
-            OnStartEntrance();
             var player = PlayerController.Instance;
-            player.transform.DOMove(entrancePoint.position, entranceTime)
-                            .SetDelay(entranceWaitTime)
-                            .OnComplete(
-                                () =>
-                                {
-                                    OnEndEntrance();
-                                }
-                            );
-        }
-
-        //pre: PlayerController.Instance != null
-        //post: sets player stats for entrance
-        private void OnStartEntrance()
-        {
-            var player = PlayerController.Instance;
-            //facing player
-            PlayerFacing facing = spawnPoint.position.x - entrancePoint.position.x > 0
-            ? PlayerFacing.Left : PlayerFacing.Right;
-            player.SetFacing(facing);
             //positioning
             player.transform.position = spawnPoint.position;
-            //starting animation of runing
-            player.Animator.SetBool(CharacterAnimations.Running, true);
-            //while entrance, player controll disabled
+            //where to face player
+            PlayerFacing facing = spawnPoint.position.x - entrancePoint.position.x > 0
+            ? PlayerFacing.Left : PlayerFacing.Right;
+            //player not controllable 
             player.Controllable = false;
+
+            MovePlayer.Trigger(entrancePoint,entranceWaitTime,facing,entranceTime,() =>{
+                player.Controllable = true;
+            });
+
         }
 
-        //pre: PlayerController.Instance != null
-        //post: sets player stats for gameplay
-        private void OnEndEntrance()
-        {
-            var player = PlayerController.Instance;
-            //starting animation of runing
-            player.Animator.SetBool(CharacterAnimations.Running, false);
-            //entrance ended, player can move
-            player.Controllable = true;
-        }
     }
 }
