+++ Erlang-Legacy/Assets/Scripts/NPC.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5620b3b272bd73348a101be2d0773dd1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
