+++ Erlang-Legacy/Assets/Scripts/NPC/Dialogue.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 48a7469f75208a341ba56228677a1d5f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
