+++ Erlang-Legacy/Assets/Scripts/NPC/Util.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ae083a25b1c1c364a8818a69e4d73a5d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
