+++ Erlang-Legacy/Assets/Scripts/NPC/Util/NPCAnimations.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8f50e099878db8142a1f2c51f4a8bea0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
