+++ Erlang-Legacy/Assets/Scripts/NPC/Util/NPCData.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using Core.NPC.Util;
+
+
+[System.Serializable]
+public class NPCData
+{
+    public string npcName;
+    
+    [TextArea(3,135)]
+    public string[] phrases;
+
+    public NPCActions[] npcActions;
+
+}
