--- Erlang-Legacy/Assets/Scripts/NPCData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 06b952b19acf062489536393f81a31ab
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
