+++ Erlang-Legacy/Assets/Scripts/Shared/MathParabola.cs
@@ -1,25 +1,28 @@
 ﻿using System;
 using UnityEngine;
 
-public class MathParabola
+namespace Core.Shared
 {
-
-    public static Vector3 Parabola(Vector3 start, Vector3 end, float height, float t)
+    public class MathParabola
     {
-        Func<float, float> f = x => -4 * height * x * x + 4 * height * x;
 
-        var mid = Vector3.Lerp(start, end, t);
+        public static Vector3 Parabola(Vector3 start, Vector3 end, float height, float t)
+        {
+            Func<float, float> f = x => -4 * height * x * x + 4 * height * x;
 
-        return new Vector3(mid.x, f(t) + Mathf.Lerp(start.y, end.y, t), mid.z);
+            var mid = Vector3.Lerp(start, end, t);
+
+            return new Vector3(mid.x, f(t) + Mathf.Lerp(start.y, end.y, t), mid.z);
+        }
+
+        public static Vector2 Parabola(Vector2 start, Vector2 end, float height, float t)
+        {
+            Func<float, float> f = x => -4 * height * x * x + 4 * height * x;
+
+            var mid = Vector2.Lerp(start, end, t);
+
+            return new Vector2(mid.x, f(t) + Mathf.Lerp(start.y, end.y, t));
+        }
+
     }
-
-    public static Vector2 Parabola(Vector2 start, Vector2 end, float height, float t)
-    {
-        Func<float, float> f = x => -4 * height * x * x + 4 * height * x;
-
-        var mid = Vector2.Lerp(start, end, t);
-
-        return new Vector2(mid.x, f(t) + Mathf.Lerp(start.y, end.y, t));
-    }
-
 }
