+++ Erlang-Legacy/Assets/Scripts/Shared/MovePlayer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 37574d04607ebc34aa16ab55d9aa6719
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
